refactored sales navigation and pages
